chore: factor out tofuresolver
Auto-installing and resolving tofu executable is a slightly separate concern from tfsandbox so putting it into a
dedicated Go package here. This also allows using it to test more Go packages that need a tofu executable for testing,
besides tofusandbox package itself.
